All Files in an R Project Folder - INCOMPLETE
Made an R Project folder and have updated some of the files so that column names that are used in code don't have spaces in them. Also added a lot of explanation of how code works above each function (Haven't done for all the admin functions). Have also refactored some of the code but need assistance for some other parts.

This is only done in the pre ME polls R code not others yet